+++ include/codecs/dct.h
@@ -1,0 +1,8 @@
+#ifndef DCT_H_
+#define DCT_H_
+
+#include "codecs/codec.h"
+
+extern const Codec DctCodec;
+
+#endif // DCT_H_